backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
pandas
numpy
scikit-learn
matplotlib
seaborn
motor
python-dotenv
python-multipart
joblib
httpx
